Re-authenticate after an unauthorized response
If an unauthorized response is received from the server, we'll now retry the request. This handles the situation where the session is removed or a login policy is updated, and the exporter can no-longer make requests.

Unfortunately the GraphQL returns a 200 for unauthorized responses, so we need to look at the error text.

Issue: #10