Update client.go
I changed the panic statement in the URL, as the url.Parse function returns an error that also can be returned by the method. I think using errors for this purpose is better, as panicking is generally discouraged in Go as far as I know.

Also, combined err != nil checks into a single check. Used named return values to make the code more self-documenting